Backend/pyproject.toml
[project]
name = "05-01-llmops-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.0.2",
    "flask-migrate>=4.1.0",
    "flask-wtf>=1.2.2",
    "injector==0.21.0",
    "openai>=2.9.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "langchain-core>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.4.1",
    "langchain>=1.2.2",
    "qianfan>=0.4.12.3",
    "langchain-huggingface>=1.2.0",
    "sentence-transformers>=5.2.0",
    "langchain-pinecone>=0.2.13",
    "faiss-cpu>=1.13.2",
    "weaviate>=0.1.2",
    "langchain-weaviate>=0.0.6",
    "unstructured>=0.18.27",
    "msoffcrypto-tool>=0.4.0",
    "markdown>=3.7",
    "pandas>=3.0.0",
    "openpyxl>=3.0.0",
    "python-docx>=1.1.0",
    "python-pptx>=1.0.0",
    "langchain-experimental>=0.4.1",
    "jieba>=0.42.1",
    "langchain-nvidia-ai-endpoints>=1.0.2",
    "langchain-classic>=1.0.1",
    "rank-bm25>=0.2.2",
    "lark>=1.3.1",
    "ddgs>=9.10.0",
    "langchain-cohere>=0.5.0",
    "dashscope>=1.25.12",
]
